# smart-backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  handlers.LoginData:
    properties:
      access_expires_in:
        example: 900
        type: integer
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      phone:
        example: "+77771234567"
        type: string
      refresh_expires_in:
        example: 2592000
        type: integer
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - access_expires_in
    - access_token
    - id
    - phone
    - refresh_expires_in
    - refresh_token
    type: object
  handlers.LoginRequest:
    properties:
      password:
        example: password123
        type: string
      phone:
        example: "+77771234567"
        type: string
    required:
    - password
    - phone
    type: object
  handlers.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.LoginData'
      message:
        example: Login successful
        type: string
      success:
        example: true
        type: boolean
    required:
    - message
    - success
    type: object
...
  handlers.PhoneAvailabilityResponse:
    properties:
      available:
        description: true if phone is available, false if already in use
        example: true
        type: boolean
      message:
        example: Phone availability checked
        type: string
      success:
        example: true
        type: boolean
    required:
    - available
    - message
    - success
    type: object
  handlers.RefreshData:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - access_token
    type: object
  handlers.RefreshRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  handlers.RefreshResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.RefreshData'
      message:
        example: Token refreshed successfully
        type: string
      success:
        example: true
        type: boolean
    required:
    - message
    - success
    type: object
  handlers.RegisterData:
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      phone:
        example: "+77771234567"
        type: string
    required:
    - id
    - phone
    type: object
  handlers.RegisterRequest:
    properties:
      password:
        example: password123
        minLength: 6
        type: string
      phone:
        example: "+77771234567"
        type: string
    required:
    - password
    - phone
    type: object
  handlers.RegisterResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.RegisterData'
      message:
        example: User registered successfully
        type: string
      success:
        example: true
        type: boolean
    required:
    - message
    - success
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ololo-gate.com
    name: API Support
  description: Secure phone-based authentication backend for Ololo Gate management
    system with dual authentication (users & admins) and role-based access control.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Ololo Gate API
  version: "1.0"
paths:
...
  /api/v1/auth/check-phone:
    get:
      consumes:
      - application/json
      description: Check if a phone number is available for registration or account
        creation (public endpoint, no authentication required)
      parameters:
      - description: Phone number in E.164 format (e.g., +77771234567)
        in: query
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Phone availability check result
          schema:
            $ref: '#/definitions/handlers.PhoneAvailabilityResponse'
        "400":
          description: Invalid phone number format
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Check if phone number is available for registration
      tags:
      - User Authentication
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with phone and password, returns access and refresh
        tokens. Supports device-based token invalidation.
      parameters:
      - description: Unique device identifier (optional - if provided and different
          from current device, previous tokens will be invalidated)
        in: query
        name: device_id
        type: string
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with tokens
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Invalid request body or phone format
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: User login
      tags:
      - User Authentication
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new access token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New access token generated
          schema:
            $ref: '#/definitions/handlers.RefreshResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Invalid or expired refresh token, or token has been invalidated
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Refresh access token
      tags:
      - User Authentication
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account with phone number and password (E.164
        format required)
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/handlers.RegisterResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "409":
          description: User with this phone number already exists
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Register a new user
      tags:
      - User Authentication
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"